+++ annotations/nocedit.exe/pseudocode/src/core/drone.cpp/FUN_0048f360.cpp
@@ -25,8 +25,8 @@ void __cdecl core_drone_cpp_FUN_0048f360(void)
             (*(uint *)(in_stack_00000004[1].base.base.actor_name + 0x1c));
   sound_sndmain_cpp_killSfx_FUN_005a9c40
             (*(uint *)(in_stack_00000004[1].base.base.actor_name + 0x18));
-  fVar1 = (float)(in_stack_00000004->base).hit_points - *(float *)(in_stack_00000008 + 4);
-  (in_stack_00000004->base).hit_points = (int)fVar1;
+  fVar1 = (in_stack_00000004->base).hit_points - *(float *)(in_stack_00000008 + 4);
+  (in_stack_00000004->base).hit_points = fVar1;
   if (0.0 < fVar1) {
     sound_sndmain_cpp_killSfx_FUN_005a9c40
               (*(uint *)(in_stack_00000004[1].base.base.actor_name + 0x18));
@@ -55,7 +55,7 @@ void __cdecl core_drone_cpp_FUN_0048f360(void)
   }
   else {
     this_ptr = &(in_stack_00000004->base).model;
-    (in_stack_00000004->base).hit_points = 0;
+    (in_stack_00000004->base).hit_points = 0.0;
     pSVar2 = core_motion_cpp_CMotionController_getCurrentMotion_FUN_0052dab0
                        (&this_ptr->motion_controller);
     if ((pSVar2->state_index != 8) && (pSVar2->state_index != 7)) {
