+++ annotations/nocedit.exe/pseudocode/src/core/imp.cpp/CImp_processDamage_FUN_004fae00.cpp
@@ -29,9 +29,9 @@ void __cdecl core_imp_cpp_CImp_processDamage_FUN_004fae00(CImp *this_ptr,SDamage
     } while (iVar4 < 5);
   }
   core_imp_cpp_CImp_FUN_004fab60(this_ptr);
-  fVar1 = (float)(this_ptr->base).base.hit_points - damage_info->damage_amount;
+  fVar1 = (this_ptr->base).base.hit_points - damage_info->damage_amount;
   this_ptr_00 = &(this_ptr->base).base.model;
-  (this_ptr->base).base.hit_points = (int)fVar1;
+  (this_ptr->base).base.hit_points = fVar1;
   if (0.0 < fVar1) {
     core_motion_cpp_CMotionController_setDesiredState_FUN_0052db00
               (&this_ptr_00->motion_controller,5,1);
@@ -40,7 +40,7 @@ void __cdecl core_imp_cpp_CImp_processDamage_FUN_004fae00(CImp *this_ptr,SDamage
     sound_name = "imp-hurt?.wav";
   }
   else {
-    (this_ptr->base).base.hit_points = 0;
+    (this_ptr->base).base.hit_points = 0.0;
     pSVar2 = core_motion_cpp_CMotionController_getCurrentMotion_FUN_0052dab0
                        (&this_ptr_00->motion_controller);
     if ((pSVar2->state_index == 6) || (pSVar2->state_index == 7)) goto LAB_004faefc;
